move `waitForAnimations` to `fixtures.js`, make a method private
