+++ e2e/tests/plugins/imagery/exampleImagery.e2e.spec.js
@@ -26,9 +26,8 @@ but only assume that example imagery is present.
 */
 /* globals process */
 
-const { test } = require('../../../fixtures.js');
+const { test, waitForAnimations } = require('../../../fixtures.js');
 const { expect } = require('@playwright/test');
-const { waitForAnimations } = require('../../../appActions.js');
 
 const backgroundImageSelector = '.c-imagery__main-image__background-image';
